+++ examples/logs/outputs.tf
@@ -8,11 +8,6 @@ output "tamr_es_domain_endpoint" {
   description = "Endpoint of ES domain created"
 }
 
-output "es_security_group_id" {
-  value       = module.tamr-es-cluster.es_security_group_ids
-  description = "ID of the ES domain created"
-}
-
 output "vpc_id" {
   value = aws_vpc.es_vpc.id
 }