+++ variables.tf
@@ -1,5 +1,4 @@
 variable "domain_name" {
-  default     = "tamr-es-cluster"
   type        = string
   description = "The name to give to the OpenSearch domain"
 }
@@ -78,57 +77,9 @@ variable "tags" {
   default     = {}
 }
 
-variable "aws_region" {
-  default     = "us-east-1"
-  type        = string
-  description = "AWS region to launch in"
-}
-
 variable "vpc_id" {
   type        = string
   description = "The ID of the VPC in which to attach the security group"
-}
-
-variable "sg_name" {
-  type        = string
-  description = "Security Group to create"
-  default     = "es-security-group"
-}
-
-variable "revoke_rules_on_delete" {
-  type        = bool
-  description = "Whether to revoke rules from the SG upon deletion"
-  default     = true
-}
-
-variable "enable_https" {
-  type        = bool
-  description = "If set to true, enables SSH"
-  default     = true
-}
-
-variable "enable_http" {
-  type        = bool
-  description = "If set to true, enables SSH"
-  default     = true
-}
-
-variable "ingress_cidr_blocks" {
-  type        = list(string)
-  description = "CIDR blocks to attach to security groups for ingress"
-  default     = []
-}
-
-variable "ingress_security_groups" {
-  type        = list(string)
-  description = "Existing security groups to attach to new security groups for ingress"
-  default     = []
-}
-
-variable "sg_tags" {
-  type        = map(string)
-  description = "Additional tags to be attached to the security group"
-  default     = {}
 }
 
 variable "kms_key_id" {
@@ -145,6 +96,7 @@ variable "enforce_https" {
   type        = bool
   description = "Whether or not to require HTTPS on the domain endpoint"
 }
+
 variable "tls_security_policy" {
   default     = "Policy-Min-TLS-1-2-2019-07"
   type        = string